pyproject.toml
[tool.poetry]
name = "international-delivery-project"
version = "0.1.0"
description = ""
authors = [""]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi-slim = "^0.112.0"
uvicorn = {extras = ["standart"], version = "^0.30.5"}
aiomysql = "^0.2.0"
sqlalchemy = "^2.0.32"
cryptography = "^43.0.0"
loguru = "^0.7.2"
httpx = "^0.27.0"
redis = "^5.0.8"
aio-pika = "^9.4.2"
jinja2 = "^3.1.4"
itsdangerous = "^2.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
